tests/load-tests/requirements.txt
locustio==0.14.6
flask==1.1.2
pyyaml
sentry-sdk==0.14.3
